requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
playwright>=1.49.0
flask>=3.0.0
